fix : 거리 재계산

See merge request s09-final/S09P31A403!34

+++ backend/src/main/java/com/ssafy/backend/dto/HospitalInfoDto.java
@@ -29,7 +29,7 @@ public class HospitalInfoDto {
 
 	private Integer minPrice;
 
-	private Double distance;
+	private Long distance;
 
 	private String treatmentName;
 }
